requirements.txt
-i https://pypi.org/simple
absl-py==2.3.1; python_version >= '3.8'
accelerate==1.10.1; python_full_version >= '3.9.0'
aiohappyeyeballs==2.6.1; python_version >= '3.9'
aiohttp==3.12.15; python_version >= '3.9'
aiosignal==1.4.0; python_version >= '3.9'
altair==5.5.0; python_version >= '3.9'
annotated-types==0.7.0; python_version >= '3.8'
asttokens==3.0.0; python_version >= '3.8'
attrs==25.3.0; python_version >= '3.8'
bert-score==0.3.13; python_version >= '3.6'
blinker==1.9.0; python_version >= '3.9'
blis==1.3.0; python_version < '3.14' and python_version >= '3.6'
cachetools==6.2.0; python_version >= '3.9'
casefy==1.1.0; python_version >= '3.6'
catalogue==2.0.10; python_version >= '3.6'
certifi==2025.8.3; python_version >= '3.7'
charset-normalizer==3.4.3; python_version >= '3.7'
click==8.2.1; python_version >= '3.10'
cloudpathlib==0.22.0; python_version >= '3.9'
cloudpickle==3.1.1; python_version >= '3.8'
coloraide==5.1; python_version >= '3.9'
confection==0.1.5; python_version >= '3.6'
contourpy==1.3.3; python_version >= '3.11'
cycler==0.12.1; python_version >= '3.8'
cymem==2.0.11
datasets==4.0.0; python_full_version >= '3.9.0'
decorator==5.2.1; python_version >= '3.8'
dill==0.3.8; python_version >= '3.8'
https://github.com/explosion/spacy-models/releases/download/en_core_web_md-3.8.0/en_core_web_md-3.8.0-py3-none-any.whl
evaluate==0.4.5; python_full_version >= '3.8.0'
executing==2.2.1; python_version >= '3.8'
filelock==3.19.1; python_version >= '3.9'
fonttools==4.59.2; python_version >= '3.9'
frozenlist==1.7.0; python_version >= '3.9'
fsspec[http]==2025.3.0; python_version >= '3.8'
gitdb==4.0.12; python_version >= '3.7'
gitpython==3.1.45; python_version >= '3.7'
grpcio==1.74.0; python_version >= '3.9'
hf-xet==1.1.9; python_version >= '3.8'
huggingface-hub==0.34.4; python_full_version >= '3.8.0'
idna==3.10; python_version >= '3.6'
imageio==2.37.0; python_version >= '3.9'
ipython==9.5.0; python_version >= '3.11'
ipython-pygments-lexers==1.1.1; python_version >= '3.8'
jedi==0.19.2; python_version >= '3.6'
jinja2==3.1.6; python_version >= '3.7'
joblib==1.5.2; python_version >= '3.9'
jsonschema==4.25.1; python_version >= '3.9'
jsonschema-specifications==2025.9.1; python_version >= '3.9'
kiwisolver==1.4.9; python_version >= '3.10'
langcodes==3.5.0; python_version >= '3.9'
language-data==1.3.0
lazy-loader==0.4; python_version >= '3.7'
lime==0.2.0.1; python_version >= '3.5'
llvmlite==0.44.0; python_version >= '3.10'
marisa-trie==1.3.1; python_version >= '3.9'
markdown==3.9; python_version >= '3.9'
markdown-it-py==4.0.0; python_version >= '3.10'
markupsafe==3.0.2; python_version >= '3.9'
matplotlib==3.10.6; python_version >= '3.10'
matplotlib-inline==0.1.7; python_version >= '3.8'
mdurl==0.1.2; python_version >= '3.7'
mpmath==1.3.0
multidict==6.6.4; python_version >= '3.9'
multiprocess==0.70.16; python_version >= '3.8'
murmurhash==1.0.13; python_version < '3.14' and python_version >= '3.6'
narwhals==2.4.0; python_version >= '3.9'
networkx==3.5; python_version >= '3.11'
nltk==3.9.1; python_version >= '3.8'
numba==0.61.2; python_version >= '3.10'
numpy==2.2.6; python_version >= '3.10'
nvidia-cublas-cu12==12.8.4.1; python_version >= '3'
nvidia-cuda-cupti-cu12==12.8.90; python_version >= '3'
nvidia-cuda-nvrtc-cu12==12.8.93; python_version >= '3'
nvidia-cuda-runtime-cu12==12.8.90; python_version >= '3'
nvidia-cudnn-cu12==9.10.2.21; python_version >= '3'
nvidia-cufft-cu12==11.3.3.83; python_version >= '3'
nvidia-cufile-cu12==1.13.1.3; python_version >= '3'
nvidia-curand-cu12==10.3.9.90; python_version >= '3'
nvidia-cusolver-cu12==11.7.3.90; python_version >= '3'
nvidia-cusparse-cu12==12.5.8.93; python_version >= '3'
nvidia-cusparselt-cu12==0.7.1; platform_system == 'Linux' and platform_machine == 'x86_64'
nvidia-nccl-cu12==2.27.3; python_version >= '3'
nvidia-nvjitlink-cu12==12.8.93; python_version >= '3'
nvidia-nvtx-cu12==12.8.90; python_version >= '3'
packaging==25.0; python_version >= '3.8'
pandas==2.3.2; python_version >= '3.9'
parso==0.8.5; python_version >= '3.6'
pexpect==4.9.0; sys_platform != 'win32' and sys_platform != 'emscripten'
pillow==11.3.0; python_version >= '3.9'
preshed==3.0.10; python_version < '3.14' and python_version >= '3.6'
prompt-toolkit==3.0.52; python_version >= '3.8'
propcache==0.3.2; python_version >= '3.9'
protobuf==6.32.0; python_version >= '3.9'
psutil==7.0.0; python_version >= '3.6'
ptyprocess==0.7.0
pure-eval==0.2.3
pyarrow==21.0.0; python_version >= '3.9'
pydantic==2.11.7; python_version >= '3.9'
pydantic-core==2.33.2; python_version >= '3.9'
pydantic-settings==2.10.1; python_version >= '3.9'
pydeck==0.9.1; python_version >= '3.8'
pygments==2.19.2; python_version >= '3.8'
pyparsing==3.2.3; python_version >= '3.9'
python-dateutil==2.9.0.post0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
python-dotenv==1.1.1; python_version >= '3.9'
pytz==2025.2
pyyaml==6.0.2; python_version >= '3.8'
referencing==0.36.2; python_version >= '3.9'
regex==2025.9.1; python_version >= '3.9'
requests==2.32.5; python_version >= '3.9'
rich==14.1.0; python_full_version >= '3.8.0'
rpds-py==0.27.1; python_version >= '3.9'
safetensors==0.6.2; python_version >= '3.9'
scikit-image==0.25.2; python_version >= '3.10'
scikit-learn==1.7.1; python_version >= '3.10'
scipy==1.16.1; python_version >= '3.11'
sentence-transformers==5.1.0; python_version >= '3.9'
seqeval==1.2.2
setuptools==80.9.0; python_version >= '3.9'
shap==0.48.0; python_version >= '3.9'
shellingham==1.5.4; python_version >= '3.7'
six==1.17.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
slicer==0.0.8; python_version >= '3.6'
smart-open==7.3.1; python_version >= '3.8' and python_version < '4.0'
smmap==5.0.2; python_version >= '3.7'
spacy==3.8.7; python_version < '3.14' and python_version >= '3.9'
spacy-legacy==3.0.12; python_version >= '3.6'
spacy-loggers==1.0.5; python_version >= '3.6'
srsly==2.5.1; python_version < '3.14' and python_version >= '3.9'
stack-data==0.6.3
streamlit==1.49.1; python_version >= '3.9' and python_full_version != '3.9.7'
sympy==1.14.0; python_version >= '3.9'
tenacity==9.1.2; python_version >= '3.9'
tensorboard==2.20.0; python_version >= '3.9'
tensorboard-data-server==0.7.2; python_version >= '3.7'
thinc==8.3.6; python_version < '3.14' and python_version >= '3.9'
threadpoolctl==3.6.0; python_version >= '3.9'
tifffile==2025.8.28; python_version >= '3.11'
tokenizers==0.22.0; python_version >= '3.9'
toml==0.10.2; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2'
torch==2.8.0; python_full_version >= '3.9.0'
tornado==6.5.2; python_version >= '3.9'
tqdm==4.67.1; python_version >= '3.7'
traitlets==5.14.3; python_version >= '3.8'
transformers==4.56.1; python_full_version >= '3.9.0'
triton==3.4.0; python_version < '3.14' and python_version >= '3.9'
typer==0.17.4; python_version >= '3.7'
typing-extensions==4.15.0; python_version >= '3.9'
typing-inspection==0.4.1; python_version >= '3.9'
tzdata==2025.2; python_version >= '2'
urllib3==2.5.0; python_version >= '3.9'
wasabi==1.1.3; python_version >= '3.6'
watchdog==6.0.0; python_version >= '3.9'
wcwidth==0.2.13
weasel==0.4.1; python_version >= '3.7'
werkzeug==3.1.3; python_version >= '3.9'
wordcloud==1.9.4; python_version >= '3.7'
wrapt==1.17.3; python_version >= '3.8'
xformers==0.0.32.post2; python_version >= '3.9'
xxhash==3.5.0; python_version >= '3.7'
yarl==1.20.1; python_version >= '3.9'
